fing height

+++ 9-binary_tree_height.c
@@ -12,8 +12,8 @@ size_t binary_tree_height(const binary_tree_t *tree)
 {
 	if (!tree)
 		return (0);
-	int left_height = binary_tree_height(tree->left);
-	int right_height = binary_tree_height(tree->right);
+	size_t left_height = binary_tree_height(tree->left);
+	size_t right_height = binary_tree_height(tree->right);
 
 	return (1 + (right_height > left_height ? right_height : left_height));
 }
